--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -1,4 +1,5 @@
-// +build !bench
+//go:build bench
+// +build bench
 
 package hw10programoptimization